Move FormTabs to flat structure
