+++ packages/components/src/components/FormTabs/FormTabs.stories.tsx
