+++ packages/components/src/components/FormTabs/index.ts
@@ -9,14 +9,4 @@
  * Copyright (C) The OpenCRVS Authors. OpenCRVS and the OpenCRVS
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
-export * from './InputField/InputField'
-export * from './InputField/InputError'
-export * from './TextInput'
-export * from './RadioGroup/RadioGroup'
-export * from './TextArea'
-export * from './SubSectionDivider'
-export * from './VerifyingIndicator'
-export * from './ValidIndicator'
-export * from './CheckboxGroup/CheckboxGroup'
-export * from './PDFViewer/PDFViewer'
-export * from './InputField/InputLabel'
+export * from './FormTabs'
